Add access control to get/create/update/delete channels.

+++ routes/api/channels.js
@@ -6,13 +6,21 @@ const ACL = require('../../acl/property_access');
 const Channel = require('../../models/Channel');
 
 router.get("/allchannels", (req, res) => {
-  Channel.find({}, 'name type link')
-    .then(channels => res.json(channels))
-    .catch(
-      err => res.status(400).json({
-        ok: false,
-        message: err
-      }));
+  const userId = req.user.id;
+  const propertyType = "channel";
+  const role = AclConfig.StrictRoles.READ_DETAIL;
+  ACL.checkAccess(userId, "", propertyType, role, (err, isAccessible) => {
+    if (!isAccessible) {
+      return res.status(401).json({err: 'No permission to read channel.'});
+    }
+    Channel.find({}, 'name type link')
+      .then(channels => res.json(channels))
+      .catch(
+        err => res.status(400).json({
+          ok: false,
+          message: err
+        }));
+  });
 });
 
 router.post("/newchannel", [
@@ -23,30 +31,38 @@ router.post("/newchannel", [
 	if (!errors.isEmpty()) {
 		return res.status(422).json({ error: errors.array() });
 	}
-  Channel
-    .findOne({name: req.body.name, type: req.body.type})
-    .then(channel => {
-      if (channel) {
-        res.status(400).json(
-          {ok: false, message: "Channel already exists"});
-      } else {
-        new Channel({
-          name: req.body.name,
-          type: req.body.type,
-          link: req.body.link,
-          auth: req.body.auth,
-          updated_at: new Date()
-        })
-        .save()
-        .then(saved => res.json({ok: true}))
-        .catch(err => {
-          res.status(400).json({ok: false, message: JSON.stringify(err)});
-        })
-      }
-    })
-    .catch(err => {
-      res.status(400).json({ok: false, message: "Failed to query database"});
-    });
+  const userId = req.user.id;
+  const propertyType = "channel";
+  const role = AclConfig.StrictRoles.ADMIN;
+  ACL.checkAccess(userId, "", propertyType, role, (err, isAccessible) => {
+    if (!isAccessible) {
+      return res.status(401).json({err: 'No permission to create channel.'});
+    }
+    Channel
+      .findOne({name: req.body.name, type: req.body.type})
+      .then(channel => {
+        if (channel) {
+          res.status(400).json(
+            {ok: false, message: "Channel already exists"});
+        } else {
+          new Channel({
+            name: req.body.name,
+            type: req.body.type,
+            link: req.body.link,
+            auth: req.body.auth,
+            updated_at: new Date()
+          })
+          .save()
+          .then(saved => res.json({ok: true}))
+          .catch(err => {
+            res.status(400).json({ok: false, message: JSON.stringify(err)});
+          })
+        }
+      })
+      .catch(err => {
+        res.status(400).json({ok: false, message: "Failed to query database"});
+      });
+  });
 });
 
 router.post("/updatechannel", [
@@ -57,19 +73,27 @@ router.post("/updatechannel", [
 	if (!errors.isEmpty()) {
 		return res.status(422).json({ error: errors.array() });
 	}
-  Channel
-    .findByIdAndUpdate(req.body._id, req.body.update)
-    .then(channel => {
-      if (!channel) {
-        res.status(400).json({ok: false, message: "Channel not found"});
-      } else {
-        res.json({ok: true});
-      }
-    })
-    .catch(err => {
-      res.status(400).json(
-        {ok: false, message: "Failed to update the channel"});
-    });
+  const userId = req.user.id;
+  const propertyType = "channel";
+  const role = AclConfig.StrictRoles.READ_WRITE;
+  ACL.checkAccess(userId, "", propertyType, role, (err, isAccessible) => {
+    if (!isAccessible) {
+      return res.status(401).json({err: 'No permission to update channel.'});
+    }
+    Channel
+      .findByIdAndUpdate(req.body._id, req.body.update)
+      .then(channel => {
+        if (!channel) {
+          res.status(400).json({ok: false, message: "Channel not found"});
+        } else {
+          res.json({ok: true});
+        }
+      })
+      .catch(err => {
+        res.status(400).json(
+          {ok: false, message: "Failed to update the channel"});
+      });
+  });
 });
 
 router.post("/deletechannel", [
@@ -79,13 +103,20 @@ router.post("/deletechannel", [
 	if (!errors.isEmpty()) {
 		return res.status(422).json({ error: errors.array() });
 	}
-  Channel
-    .findByIdAndRemove(req.body._id)
-    .then(data => {
-      res.json({ok: true, data: data});
-    }).catch((err) => {
-      res.status(400).json({"errReason": err});
-    });
+  const userId = req.user.id;
+  const propertyType = "channel";
+  ACL.checkAccess(userId, "", propertyType, "owner", (err, isAccessible) => {
+    if (!isAccessible) {
+      return res.status(401).json({err: 'No permission to delete channel.'});
+    }
+    Channel
+      .findByIdAndRemove(req.body._id)
+      .then(data => {
+        res.json({ok: true, data: data});
+      }).catch((err) => {
+        res.status(400).json({"errReason": err});
+      });
+  });
 });
 
 module.exports = router;
